+++ resources/views/admin/dashboard/index.php
@@ -7,6 +7,7 @@ declare(strict_types=1);
  * @var View $this
  */
 
+use Hirtz\Skeleton\Modules\Admin\Widgets\EnvironmentAlert;
 use Hirtz\Skeleton\Modules\Admin\Widgets\Navs\DashboardHeader;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Panels\Dashboard;
@@ -14,4 +15,5 @@ use Hirtz\Skeleton\Widgets\Panels\Dashboard;
 $this->title(Yii::t('skeleton', 'DASHBOARD_NAV_ITEM_DASHBOARD'));
 
 echo DashboardHeader::make();
+echo EnvironmentAlert::make();
 echo Dashboard::make();